Transferred over the existing UI from Jarvis. Changes will still need
to be made. First the back end needs to be updated to include the
database and model for the User so that we can create an account
and login. Jarvis controlled a full house so the logic for alfred is
slightly different. As you can see A LOT of work has already been
implemented and there is no reason to keep reinventing the wheel. We
will repurpose it here.