Adjusted the existing Jarvis implementation and refactored for one room.
All of the previous work comes from another private repository,
jarvis_home. Which controlled lights and tvs across different rooms.

Alfred, is Jarvis reincarnated which controls wemo outlets and a smart
tv in one room.

Currently, pre-authorized users can create an account, and control the
wemo devices. Further testing needed to be done but I have a ton of
changes that are not pushed to the repo. It would have been silly to
make multiple unnecessary commits as all this needed to be addressed to
have the UI back into a working condition.